GoodボタンとStayボタンのAPIイベントを親側で統一
